arrowfunctions and moved functions to resources
